smart_kitchen/requirements.txt
fastapi
uvicorn
pandas
pydantic
requests
streamlit
matplotlib
scikit-learn
paho-mqtt
flask
streamlit-authenticator
python-dotenv